admin_service/requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.22.0
SQLAlchemy==2.0.22
psycopg2-binary==2.9.6
jinja2==3.1.2
python-dotenv==1.0.0
requests==2.31.0
